backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
httpx==0.28.1
pdfplumber==0.11.4
beautifulsoup4==4.12.3
lxml==5.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic[email]>=2.0.0
python-multipart>=0.0.9
